+++ core/src/main/java/org/vertexium/VertexiumSerializer.java
@@ -1,0 +1,7 @@
+package org.vertexium;
+
+public interface VertexiumSerializer {
+    byte[] objectToBytes(Object object);
+
+    <T> T bytesToObject(byte[] bytes);
+}