+++ kryo-serializer/src/main/java/org/vertexium/serializer/kryo/quickSerializers/BigDecimalQuickTypeSerializer.java
@@ -1,18 +1,16 @@
-package org.vertexium.accumulo.serializer.kryo.quickSerializers;
-
-import org.apache.accumulo.core.data.Value;
+package org.vertexium.serializer.kryo.quickSerializers;
 
 import java.math.BigDecimal;
 
 public class BigDecimalQuickTypeSerializer implements QuickTypeSerializer {
     @Override
-    public Value objectToValue(Object value) {
+    public byte[] objectToBytes(Object value) {
         BigDecimal valueAsBigDecimal = (BigDecimal) value;
         byte[] valueBytes = valueAsBigDecimal.toString().getBytes();
         byte[] data = new byte[1 + valueBytes.length];
         data[0] = MARKER_BIG_DECIMAL;
         System.arraycopy(valueBytes, 0, data, 1, valueBytes.length);
-        return new Value(data);
+        return data;
     }
 
     @Override